# Configurable read buffer