The default read buffer size is standard for disk access, but we are using this library with object storage backing the parquet.File. With object storage hundreds of tiny reads of 4096 is obviously a non-starter so we built a [simple buffer](https://github.com/grafana/tempo/blob/main/pkg/io/buffered.go). Unfortunately this buffer often overreads by quite a bit and for performance sensitive accesses we have found that simply increasing `defaultBufferSize` in this library beats our buffer in nearly every case.

This PR makes the `defaultBufferSize` configurable to allow those who are backing their `parquet.File`s with unconventional storage some control over how it is accessed. It also removes `defaultLevelBufferSize` because it was unused. No issue with re-adding if it's there for a reason I don't understand.

Finally, I was unable to write a convincing test for this configuration option. Any advice on how to proceed would be appreciated.