@joe-elliott 2022-08-30
Added a RWMutex protected map from buffer sizes -> sync.Pool. Using the following benchmark:
```
func BenchmarkBufioGetPut(b *testing.B) {
	for i := 0; i < b.N; i++ {
		rdr := getBufioReader(nil, 100)
		putBufioReader(rdr)
	}
}
```

Compared to main (sync.Map was even slower):
```
$ benchstat before.txt after.txt 
name           old time/op    new time/op    delta
BufioGetPut-8    29.2ns ± 4%    59.0ns ± 2%  +101.81%  (p=0.000 n=10+10)

name           old alloc/op   new alloc/op   delta
BufioGetPut-8     0.00B          0.00B           ~     (all equal)

name           old allocs/op  new allocs/op  delta
BufioGetPut-8      0.00           0.00           ~     (all equal)
```

I would like to write tests against these methods but all of the testing is currently in external packages. Should I export these methods for testing? or just skip it?